scraper/requirements-min.txt
# Dependências mínimas para teste básico
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dateutil>=2.8.0

# Opcionais (comentados para teste básico)
# loguru>=0.7.0
# nltk>=3.8.0
# spacy>=3.7.0
# selenium>=4.15.0
# google-api-python-client>=2.100.0
# apscheduler>=3.10.0